permissions/geolocation.http.html

<!doctype html>
<meta charset=utf-8>
<title>Test "geolocation" Permission in HTTP context.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>

<script>
promise_test(async () => {
  const result = await navigator.permissions.query({
    name: "geolocation",
  });
  assert_equals(result.state, "denied", "Geolocation is not available in insecure contexts");
}, "getting permission status of geolocation in insecure contexts");
</script>
